fix: jwt関連の処理修正
Changes:
1. JwtAuthenticationFilter#doFilterInternal を ifPresent による簡潔な記述に変更

2. JWT 検証失敗時に throw せず 401 Unauthorized を返すよう修正

3. shouldNotFilter を /api/auth/ 全体にマッチするよう変更

4. JwtProperties の accessTokenValidityInMinutes および refreshTokenValidityInDays の型を double に変更

5. JwtProvider#createToken の有効期限計算ロジックを double 対応に変更

Reason:
1. コードの可読性・簡潔性を高めるため

2. トークンが不正な場合に例外を外へ伝播させず、クライアントに適切なレスポンスを返すため

3. リフレッシュトークン以外にも /api/auth/ 以下全体でフィルターをスキップしたいため

4. 設定ファイル上で有効期限に小数（例：0.1分など）を許容するため（long では小数値が切り捨てられ意図しない動作となる）

5. accessTokenValidityInMinutes に小数値を設定しても破綻しないよう対応するため

BREAKING CHANGE: N/A
Related: N/A
Refs: N/A

+++ backend/src/main/java/com/example/keirekipro/presentation/security/jwt/JwtProperties.java
@@ -23,10 +23,10 @@ public class JwtProperties {
     /**
      * アクセストークンの有効期限
      */
-    private long accessTokenValidityInMinutes;
+    private double accessTokenValidityInMinutes;
 
     /**
      * リフレッシュトークンの有効期限
      */
-    private long refreshTokenValidityInDays;
+    private double refreshTokenValidityInDays;
 }
